% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_sumstats_bimr_varbeta}
\alias{generate_sumstats_bimr_varbeta}
\title{Generates sumstats from our bi-directional Mendelian randomization model}
\usage{
generate_sumstats_bimr_varbeta(
  Na,
  Nb,
  M,
  sa,
  sb,
  su,
  nu,
  eta,
  gamma,
  delta,
  p,
  q,
  r,
  s,
  af_ref_file = NULL
)
}
\arguments{
\item{Na}{Number of samples in each study of phenotype A.}

\item{Nb}{Number of samples in each study of phenotype B.}

\item{M}{Total number of SNPs to simulate, including null SNPs.}

\item{sa}{Variance of the sampling distribution of SNPs effecting A.}

\item{sb}{Variance of the sampling distribution of SNPs effecting B.}

\item{su}{Variance of the sampling distribution of SNPs effecting U.}

\item{nu}{Causal effect of U on A.}

\item{eta}{Causal effect of U on B.}

\item{gamma}{Causal effect of A on B.}

\item{delta}{Causal effect of B on A.}

\item{p}{Proportion of non-null SNPs.}

\item{q}{Proportion of non-null SNPs effecting U.}

\item{r}{Proportion of non-null SNPs effecting A.}

\item{s}{Proportion of non-null SNPs effecting B.}

\item{af_ref_file}{String giving the location of a gzipped tsv with a column
called `MAF`, giving a list of reference allele frequencies to use in
simulation. Usually just an HM3 LD-score file for a particular chromosome.}
}
\description{
Generates sumstats from our bi-directional Mendelian randomization model
}
