% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{h2_to_s2}
\alias{h2_to_s2}
\title{Helper function for turning the h2 parameterization into the variance one.}
\usage{
h2_to_s2(M, ha, hb, hu, nu, eta, gamma, delta, p, q, r, s)
}
\arguments{
\item{M}{Total number of SNPs to simulate, including null SNPs.}

\item{ha}{Heritability of phenotype A.}

\item{hb}{Heritability of phenotype B.}

\item{hu}{Heritability of phenotype U.}

\item{nu}{Causal effect of U on A.}

\item{eta}{Causal effect of U on B.}

\item{gamma}{Causal effect of A on B.}

\item{delta}{Causal effect of B on A.}

\item{p}{Proportion of non-null SNPs.}

\item{q}{Proportion of non-null SNPs effecting U.}

\item{r}{Proportion of non-null SNPs effecting A.}

\item{s}{Proportion of non-null SNPs effecting B.}
}
\description{
Helper function for turning the h2 parameterization into the variance one.
}
