% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{run_sim}
\alias{run_sim}
\title{Helper function to create a dataset and compare the results of many methods.}
\usage{
run_sim(
  name,
  iter,
  Na,
  Nb,
  M,
  ha,
  hb,
  hu,
  nu,
  eta,
  gamma,
  delta,
  p,
  q,
  r,
  s,
  af_ref_file = NULL
)
}
\arguments{
\item{name}{A name for this simulation setup.}

\item{iter}{The current iteration.}

\item{Na}{Number of samples in each study of phenotype A.}

\item{Nb}{Number of samples in each study of phenotype B.}

\item{M}{Total number of SNPs to simulate, including null SNPs.}

\item{ha}{Heritability of phenotype A.}

\item{hb}{Heritability of phenotype B.}

\item{hu}{Heritability of phenotype U.}

\item{nu}{Causal effect of U on A.}

\item{eta}{Causal effect of U on B.}

\item{gamma}{Causal effect of A on B.}

\item{delta}{Causal effect of B on A.}

\item{p}{Proportion of non-null SNPs.}

\item{q}{Proportion of non-null SNPs effecting U.}

\item{r}{Proportion of non-null SNPs effecting A.}

\item{s}{Proportion of non-null SNPs effecting B.}

\item{af_ref_file}{String giving the location of a gzipped tsv with a column
called `MAF`, giving a list of reference allele frequencies to use in
simulation. Usually just an HM3 LD-score file for a particular chromosome.}
}
\description{
Helper function to create a dataset and compare the results of many methods.
}
