% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WWER.R
\name{WWER}
\alias{WWER}
\title{Fits a regression of outcome on exposure using SE and weights as weights.}
\usage{
WWER(b_exp, b_out, se_exp, se_out, weights)
}
\arguments{
\item{b_exp}{Vector of floats. SNP effects on exposure.}

\item{b_out}{Vector of floats. SNP effects on outcome.}

\item{se_exp}{Vector of floats. Standard error of SNP effects on exposure.}

\item{se_out}{Vector of floats. Standard error of SNP effects on outcome.}

\item{weights}{Vector of floats. Welch weights to use for regression, will be
combined with the SE of the outcome to determine final weights.}
}
\description{
Fits a regression of outcome on exposure using SE and weights as weights.
}
